Backend/requirements.txt
crewai
crewai-tools
langchain-groq
python-dotenv
litellm
langchain-google-genai
psycopg2
sqlalchemy
fastapi
uvicorn[standard]
httpx
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
json_repair
chromadb
sentence-transformers